Changing count function to be capped at long instead of integer

+++ integ-test/src/test/java/org/opensearch/sql/legacy/AggregationExpressionIT.java
@@ -204,7 +204,7 @@ public class AggregationExpressionIT extends SQLIntegTestCase {
                 Index.BANK.getName()));
 
     verifySchema(
-        response, schema("birthdate", null, "timestamp"), schema("count(*)", "count", "integer"));
+        response, schema("birthdate", null, "timestamp"), schema("count(*)", "count", "long"));
     verifyDataRows(response, rows("2018-06-23 00:00:00", 1));
   }
 
@@ -222,7 +222,7 @@ public class AggregationExpressionIT extends SQLIntegTestCase {
     verifySchema(
         response,
         schema("birthdate", "birth", "timestamp"),
-        schema("count(*)", "count", "integer"));
+        schema("count(*)", "count", "long"));
     verifyDataRows(response, rows("2018-06-23 00:00:00", 1));
   }
